requirements.txt
Flask
pandas
pdfminer.six
requests
scikit-learn
gunicorn
flask_cors